requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
bitsandbytes==0.49.2
certifi==2026.2.25
click==8.3.2
fastapi==0.136.0
filelock==3.29.0
frozenlist==1.8.0
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.4.4
ollama==0.6.1
packaging==26.1
propcache==0.4.1
psutil==7.2.2
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.4.0
regex==2026.4.4
rich==15.0.0
safetensors==0.7.0
setuptools==81.0.0
shellingham==1.5.4
soupsieve==2.8.3
starlette==1.0.0
sympy==1.14.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
yarl==1.23.0
